feat(ai): 对话框逻辑抽离
